+++ functions.php
@@ -1136,6 +1136,13 @@ add_action( 'pre_get_posts', function($query){
     }
 });
 
+// Add additional media type filters
+add_filter( 'post_mime_types', function($post_mime_types){
+	$post_mime_types['model/gltf-binary'] = [ __( '3D Models' ), __( 'Manage 3D Models' ), _n_noop( 'GLB <span class="count">(%s)</span>', 'GLBs <span class="count">(%s)</span>' ) ];
+    $post_mime_types['image/svg+xml'] = [ __( 'Vector Images' ), __( 'Manage SVG Images' ), _n_noop( 'SVG <span class="count">(%s)</span>', 'SVGs <span class="count">(%s)</span>' ) ];
+	return $post_mime_types;
+});
+
 
 /////////////////////////////////////////////////
 // Admin: Add Additional values to user profiles